For InputQA, switches from properties to state for propertyTemplate and lookupConfig.
closes #692
