+++ __tests__/Utilities.test.js
@@ -5,6 +5,7 @@ import {
   isResourceWithValueTemplateRef,
   resourceToName,
   templateBoolean,
+  getLookupConfigItems,
 } from '../src/Utilities'
 
 
@@ -196,6 +197,50 @@ describe('Utilities', () => {
         label: 'http://id.loc.gov/vocabulary/mcolor/mul',
         uri: 'http://id.loc.gov/vocabulary/mcolor/mul',
       }])
+    })
+  })
+  describe('getLookupConfigItems()', () => {
+    it('returns an empty array if passed any value that fails to define a `valueConstraint.useValuesFrom` array', () => {
+      expect(getLookupConfigItems({})).toEqual([])
+    })
+
+    it('returns an empty array if passed a value with an empty `valueConstraint.useValuesFrom` array', () => {
+      const template = {
+        valueConstraint: {
+          useValuesFrom: [],
+        },
+      }
+
+      expect(getLookupConfigItems(template)).toEqual([])
+    })
+
+    it('returns an array with lookupConfig objects matching URIs in useValuesFrom array', () => {
+      const template = {
+        valueConstraint: {
+          useValuesFrom: [
+            'http://does.not.match/1',
+            'http://does.not.match/2',
+            'urn:ld4p:qa:agrovoc',
+            'https://id.loc.gov/vocabulary/mrectype',
+          ],
+        },
+      }
+
+      expect(getLookupConfigItems(template)).toEqual([
+        {
+          label: 'AGROVOC (QA)',
+          uri: 'urn:ld4p:qa:agrovoc',
+          authority: 'agrovoc_ld4l_cache',
+          subauthority: '',
+          language: 'en',
+          component: 'lookup',
+        },
+        {
+          label: 'type of recording',
+          uri: 'https://id.loc.gov/vocabulary/mrectype',
+          component: 'list',
+        },
+      ])
     })
   })
 })
